03-python-pyfmg/requirements.txt
# FortiManager API Demo - pyFMG
# Install with: pip install -r requirements.txt

pyfmg>=0.8.6.3
python-dotenv>=1.0.0
